requirements.txt
python-pptx>=0.6.21
openai>=1.0.0
python-dotenv>=1.0.0
Pillow>=10.0.0
matplotlib>=3.7.0
pandas>=2.0.0
pydantic>=2.0.0
colorama>=0.4.6
